+++ app/models/camaleon_cms/widget/main.rb
@@ -3,7 +3,10 @@ module CamaleonCms
     class Main < CamaleonCms::TermTaxonomy
       normalize_attrs(:name)
 
-      default_scope { where(taxonomy: :widget) }
+      def self.sti_name
+        'widget'
+      end
+
       # attr_accessible :excerpt, :renderer
       # name: "title"
       # description: "content for this"
@@ -12,14 +15,14 @@ module CamaleonCms
       # excerpt: string for message
       # renderer: string (path to the template for render this widget)
 
-      has_many :metas, lambda {
-                         where(object_class: 'Widget::Main')
-                       }, class_name: 'CamaleonCms::Meta', foreign_key: :objectid, dependent: :destroy
-      belongs_to :owner, class_name: CamaManager.get_user_class_name, foreign_key: :user_id
-      belongs_to :site, class_name: 'CamaleonCms::Site', foreign_key: :parent_id
+      belongs_to :owner, class_name: CamaManager.get_user_class_name.to_s, foreign_key: :user_id, inverse_of: :widgets
+      belongs_to :site, class_name: 'CamaleonCms::Site', foreign_key: :parent_id, inverse_of: :widgets
 
-      has_many :assigned, class_name: 'CamaleonCms::Widget::Assigned', foreign_key: :visibility, dependent: :destroy
+      has_many :assigned, class_name: 'CamaleonCms::Widget::Assigned', foreign_key: :visibility, dependent: :destroy,
+                          inverse_of: :widget
+
       before_save :check_excerpt
+
       def is_simple?
         status == 'simple'
       end