+++ spec/models/user_spec.rb
@@ -12,4 +12,20 @@ RSpec.describe CamaleonCms::User, type: :model do
       expect(user.email).to eql('foo@bar.com')
     end
   end
+
+  describe 'widgets association' do
+    it 'fetches widgets by user_id' do
+      user = create(:user)
+      other_site = create(:site, slug: "other-site-#{SecureRandom.hex(3)}").decorate
+      widget = CamaleonCms::Widget::Main.create!(
+        taxonomy: CamaleonCms::Widget::Main.sti_name,
+        name: 'User Widget',
+        slug: "user-widget-#{SecureRandom.hex(4)}",
+        parent_id: other_site.id,
+        user_id: user.id
+      )
+
+      expect(user.widgets).to include(widget)
+    end
+  end
 end